MAVLink: support 32-bit system IDs and payload targets
Preserve unsigned system IDs throughout vehicle selection, routing, parameters, mission transfer and RC control. Extend source and target headers independently, using TARGET32 only when a real payload target exceeds 255; retain payload broadcasts and reject unknown incompatible flags.

Cover wire fixtures, signing, target filtering and full-width routing, with a Linux GUI/SITL integration harness.

+++ Controls/DigitalSkyUI.cs
@@ -267,7 +267,7 @@ namespace MissionPlanner.Controls
 
             File.WriteAllText(destlocalfile, xmlfile);
 
-            MAVFtp ftp = new MAVFtp(MainV2.comPort, (byte)MainV2.comPort.sysidcurrent,
+            MAVFtp ftp = new MAVFtp(MainV2.comPort, MainV2.comPort.sysidcurrent,
                 (byte)MainV2.comPort.compidcurrent);
 
             ftp.UploadFile(id + ".xml", destlocalfile, null);
